+++ niftilib/nifti1_io.h
@@ -294,7 +294,7 @@ void         nifti_image_free    ( nifti_image *nim ) ;
 int          nifti_read_collapsed_image( nifti_image * nim, const int dims [8],
                                          void ** data );
 
-int          nifti_read_subregion_image( nifti_image * nim,
+int64_t          nifti_read_subregion_image( nifti_image * nim,
                                          const int *start_index, const int *region_size,
                                          void ** data );
